+++ packages/components-angular/projects/components-wrapper/CHANGELOG.md
@@ -6,6 +6,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `Text List`
+
+### Changed
+- Improve caching strategy for fonts by content-based hash
+- Improve caching strategy for marque by content-based hash
+- Dimensions and sharpness of marque
+
+## [1.4.0-rc.2] - 2020-05-06
+
+### Added
+- `Content Wrapper`
+- Description property to `p-text-field-wrapper`, `p-textarea-wrapper` and `p-select-wrapper`
+- `Link Social`
+
+### Changed
+- Improve accessibility of error and success states of form elements
+- Aria-invalid attribute of form elements if they are in error state is now managed by component 
+- Rename icon name `configure` to `configurate` (prevents breaking change compared to stable v1.3.0)
+
+### Removed
+- `safe-zone` property of `p-grid` (`Content Wrapper` should be used instead)
+
 ## [1.4.0-rc.1] - 2020-04-27
 
 ### Added